requirements.txt
Appium-Python-Client==0.22
nose==1.3.7